+++ auction/src/components/Controls/Label.js
